More work on the LL vectorizer

+++ New_LL_Code_w_Pandas.py
@@ -8,16 +8,17 @@ import numpy as np
 from tkinter.filedialog import askdirectory
 import os.path
 from scipy.stats import chi2_contingency
+import datetime
 def file_dict_builder():
     orig_dir = askdirectory(title= 'Where are your collocate DataFrames and lem_dicts located?')
     dest_dir = askdirectory(title = 'Where would you like to save the resulting files?')
     file_list = os.listdir(orig_dir)
     df_dict_dict = {}
     for file in file_list:
-        if file.endswith('Coll.pickle'):
-            df_dict_dict[file] = ''.join([file.split('.')[0], '_lem_dict.pickle'])
-    return df_dict_dict, dest_dir
-def log_like(x):
+        if file.endswith('coll.pickle'):
+            df_dict_dict[file] = ''.join([file.split('.')[0], '_dict.pickle'])
+    return df_dict_dict, dest_dir, orig_dir
+def log_like(column, row):
     #For chi2 contingency test, we need first to construct a 2x2 contingency table like the one below
     #      |    w2    |    -w2    |
     #------------------------------
@@ -35,8 +36,8 @@ def log_like(x):
     #And since c(-w2) (or -w1) is N-c(w2), we can rewrite this as N-c(w2)-c(w1,-w2).
     #And since c(w1,-w2) is c(w1)-c(w1,w2), then this can be rewritten as N-c(w2)-(c(w1)-c(w1,w2)) or N-c(w2)-c(w1)+c(w1,w2).
     #Adding c(w1,w2) back at the end corrects for the number of times that w1 appears with w2 that were already counted in c(w2)
-    c1 = lem_counts[x.name] #the count of the target word (names of the row in the df)
-    LL, p, dof, expected = chi2_contingency([[x, np.fmax(c1-x, 0)],[np.fmax(lem_counts-x,0), np.fmax(N-lem_counts-c1+x,0)]], lambda_ = 0)
+    c1 = lem_counts[row] #the count of the target word (names of the row in the df)
+    LL, p, dof, expected = chi2_contingency([[Coll_df.ix[column, row], max(c1-Coll_df.ix[column, row], 0)],[max(lem_counts[column]-Coll_df.ix[column, row],0), max(N-lem_counts[column]-c1+Coll_df.ix[column, row],0)]], lambda_ = 0)
     return LL, p
 #The above returns one value for the whole row.  I need to figure out how to get it to return a value for each element.
 #I have 2 Series (x, lem_counts) and one constant.  Can I somehow map the once series onto the other with the formulae above?  I will try tomorrow.
@@ -48,18 +49,30 @@ def counter(lem_dict_filename):
     #returns this series and N, the total number of words in the corpus
     lem_dict = pd.read_pickle(lem_dict_filename)
     new_dict = {}
-    for key, value in lem_dict.values():
-        new_dict[key] = value['count']
+    for key, value in lem_dict.items():
+        new_dict[key] = value
     lem_counts = pd.Series(new_dict)
     N = lem_counts.sum()
     return lem_counts, N
 
-file_dict, dest_dir = file_dict_builder()
+file_dict, dest_dir, orig_dir = file_dict_builder()
 for df_file, lem_file in file_dict.items():
-    LL_dest_file = ''.join([dest_dir, df_file.split('_')[0], '_LL.pickle'])
-    LL_p_dest_file = ''.join([dest_dir, df_file.split('_')[0], '_LL_p.pickle'])
-    lem_counts, N = lem_counter(lem_file)
-    Coll_df = pd.read_pickle(df_file)
-    LL_df, LL_p_df = Coll_df.apply(lambda x: log_like(x), axis = 1)
-    LL_df.to_pickle(LL_dest_file)
-    LL_p_df.to_pickle(LL_p_dest_file)
+    print('Now analyzing %s'% df_file)
+    LL_dest_file = '/'.join([dest_dir, ''.join([df_file.split('_')[0], '_LL.pickle'])])
+    LL_p_dest_file = '/'.join([dest_dir, ''.join([df_file.split('_')[0], '_LL_p.pickle'])])
+    if os.path.isfile(LL_dest_file) and os.path.isfile(LL_p_dest_file):
+        continue
+    else:
+        lem_counts, N = counter('/'.join([orig_dir, lem_file]))
+        Coll_df = pd.read_pickle('/'.join([orig_dir, df_file]))
+        LL_df = pd.DataFrame(0., index = Coll_df.index, columns = Coll_df.index)
+        LL_p_df = pd.DataFrame(0., index = Coll_df.index, columns = Coll_df.index)
+        my_counter = 0
+        for row in Coll_df.index:
+            if my_counter % 100 == 0:
+                print('Row %s of %s at %s'% (my_counter, len(Coll_df), datetime.datetime.now().time().isoformat()))
+            for column in Coll_df.index:
+                LL_df.ix[column, row], LL_p_df.ix[column, row] = log_like(column, row)
+            my_counter += 1
+        LL_df.to_pickle(LL_dest_file)
+        LL_p_df.to_pickle(LL_p_dest_file)
